local web/requirement.txt
Django==2.2.13
django-pure-pagination==0.3.0
numpy==1.19.5
opencv-python
pytz==2019.3
sqlparse==0.3.0
